Finish restructuring pool module
