+++ src/dbus_api/pool/pool_2_0/methods.rs
@@ -1,1 +1,308 @@
+use std::{collections::HashMap, path::Path};
 
+use dbus::{
+    arg::Array,
+    tree::{MTFn, MethodInfo, MethodResult},
+    Message,
+};
+use devicemapper::Sectors;
+
+use crate::{
+    dbus_api::{
+        blockdev::create_dbus_blockdev,
+        filesystem::create_dbus_filesystem,
+        types::{DbusErrorEnum, TData},
+        util::{engine_to_dbus_err_tuple, get_next_arg, msg_code_ok, msg_string_ok},
+    },
+    engine::{BlockDevTier, CreateAction, EngineAction, FilesystemUuid, PoolUuid, RenameAction},
+};
+
+pub fn create_filesystems(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    let message: &Message = m.msg;
+    let mut iter = message.iter_init();
+
+    let filesystems: Array<&str, _> = get_next_arg(&mut iter, 0)?;
+    let dbus_context = m.tree.get_data();
+
+    let object_path = m.path.get_name();
+    let return_message = message.method_return();
+    let default_return: (bool, Vec<(dbus::Path, &str)>) = (false, Vec::new());
+
+    if filesystems.count() > 1 {
+        let error_message = "only 1 filesystem per request allowed";
+        let (rc, rs) = (DbusErrorEnum::ERROR as u16, error_message);
+        return Ok(vec![return_message.append3(default_return, rc, rs)]);
+    }
+
+    let pool_path = m
+        .tree
+        .get(object_path)
+        .expect("implicit argument must be in tree");
+    let pool_uuid = get_data!(pool_path; default_return; return_message).uuid;
+
+    let mut engine = dbus_context.engine.borrow_mut();
+    let (pool_name, pool) = get_mut_pool!(engine; pool_uuid; default_return; return_message);
+
+    let result = pool.create_filesystems(
+        pool_uuid,
+        &pool_name,
+        &filesystems
+            .map(|x| (x, None))
+            .collect::<Vec<(&str, Option<Sectors>)>>(),
+    );
+
+    let infos = match result {
+        Ok(created_set) => created_set.changed(),
+        Err(err) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&err);
+            return Ok(vec![return_message.append3(default_return, rc, rs)]);
+        }
+    };
+
+    let return_value = match infos {
+        Some(ref newly_created_filesystems) => {
+            let v = newly_created_filesystems
+                .iter()
+                .map(|&(name, uuid)| {
+                    // FIXME: To avoid this expect, modify create_filesystem
+                    // so that it returns a mutable reference to the
+                    // filesystem created.
+                    (
+                        create_dbus_filesystem(
+                            dbus_context,
+                            object_path.clone(),
+                            uuid,
+                            pool.get_mut_filesystem(uuid)
+                                .expect("just inserted by create_filesystems")
+                                .1,
+                        ),
+                        name,
+                    )
+                })
+                .collect::<Vec<_>>();
+            (true, v)
+        }
+        None => default_return,
+    };
+
+    Ok(vec![return_message.append3(
+        return_value,
+        msg_code_ok(),
+        msg_string_ok(),
+    )])
+}
+
+pub fn destroy_filesystems(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    let message: &Message = m.msg;
+    let mut iter = message.iter_init();
+
+    let filesystems: Array<dbus::Path<'static>, _> = get_next_arg(&mut iter, 0)?;
+
+    let dbus_context = m.tree.get_data();
+    let object_path = m.path.get_name();
+    let return_message = message.method_return();
+    let default_return: (bool, Vec<String>) = (false, Vec::new());
+
+    let pool_path = m
+        .tree
+        .get(object_path)
+        .expect("implicit argument must be in tree");
+    let pool_uuid = get_data!(pool_path; default_return; return_message).uuid;
+
+    let mut engine = dbus_context.engine.borrow_mut();
+    let (pool_name, pool) = get_mut_pool!(engine; pool_uuid; default_return; return_message);
+
+    let filesystem_map: HashMap<FilesystemUuid, dbus::Path<'static>> = filesystems
+        .filter_map(|path| {
+            m.tree.get(&path).and_then(|op| {
+                op.get_data()
+                    .as_ref()
+                    .map(|d| (d.uuid, op.get_name().clone()))
+            })
+        })
+        .collect();
+
+    let result = pool.destroy_filesystems(
+        &pool_name,
+        &filesystem_map.keys().cloned().collect::<Vec<_>>(),
+    );
+    let msg = match result {
+        Ok(uuids) => {
+            // Only get changed values here as non-existant filesystems will have been filtered out
+            // before calling destroy_filesystems
+            let uuid_vec: Vec<String> = if let Some(ref changed_uuids) = uuids.changed() {
+                for uuid in changed_uuids {
+                    let op = filesystem_map
+                        .get(uuid)
+                        .expect("'uuids' is a subset of filesystem_map.keys()");
+                    dbus_context.actions.borrow_mut().push_remove(op, m.tree);
+                }
+                changed_uuids
+                    .iter()
+                    .map(|uuid| uuid_to_string!(uuid))
+                    .collect()
+            } else {
+                Vec::new()
+            };
+            return_message.append3((true, uuid_vec), msg_code_ok(), msg_string_ok())
+        }
+        Err(err) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&err);
+            return_message.append3(default_return, rc, rs)
+        }
+    };
+    Ok(vec![msg])
+}
+
+pub fn snapshot_filesystem(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    let message: &Message = m.msg;
+    let mut iter = message.iter_init();
+
+    let filesystem: dbus::Path<'static> = get_next_arg(&mut iter, 0)?;
+    let snapshot_name: &str = get_next_arg(&mut iter, 1)?;
+
+    let dbus_context = m.tree.get_data();
+    let object_path = m.path.get_name();
+    let return_message = message.method_return();
+    let default_return = (false, dbus::Path::default());
+
+    let pool_path = m
+        .tree
+        .get(object_path)
+        .expect("implicit argument must be in tree");
+    let pool_uuid = get_data!(pool_path; default_return; return_message).uuid;
+
+    let fs_uuid = match m.tree.get(&filesystem) {
+        Some(op) => get_data!(op; default_return; return_message).uuid,
+        None => {
+            let message = format!("no data for object path {}", filesystem);
+            let (rc, rs) = (DbusErrorEnum::NOTFOUND as u16, message);
+            return Ok(vec![return_message.append3(default_return, rc, rs)]);
+        }
+    };
+
+    let mut engine = dbus_context.engine.borrow_mut();
+    let (pool_name, pool) = get_mut_pool!(engine; pool_uuid; default_return; return_message);
+
+    let msg = match pool.snapshot_filesystem(pool_uuid, &pool_name, fs_uuid, snapshot_name) {
+        Ok(CreateAction::Created((uuid, fs))) => {
+            let fs_object_path: dbus::Path =
+                create_dbus_filesystem(dbus_context, object_path.clone(), uuid, fs);
+            return_message.append3((true, fs_object_path), msg_code_ok(), msg_string_ok())
+        }
+        Ok(CreateAction::Identity) => {
+            return_message.append3(default_return, msg_code_ok(), msg_string_ok())
+        }
+        Err(err) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&err);
+            return_message.append3(default_return, rc, rs)
+        }
+    };
+
+    Ok(vec![msg])
+}
+
+pub fn add_blockdevs(m: &MethodInfo<MTFn<TData>, TData>, tier: BlockDevTier) -> MethodResult {
+    let message: &Message = m.msg;
+    let mut iter = message.iter_init();
+
+    let devs: Array<&str, _> = get_next_arg(&mut iter, 0)?;
+
+    let dbus_context = m.tree.get_data();
+    let object_path = m.path.get_name();
+    let return_message = message.method_return();
+    let default_return: (bool, Vec<dbus::Path>) = (false, Vec::new());
+
+    let pool_path = m
+        .tree
+        .get(object_path)
+        .expect("implicit argument must be in tree");
+    let pool_uuid = get_data!(pool_path; default_return; return_message).uuid;
+
+    let mut engine = dbus_context.engine.borrow_mut();
+    let (pool_name, pool) = get_mut_pool!(engine; pool_uuid; default_return; return_message);
+
+    let blockdevs = devs.map(|x| Path::new(x)).collect::<Vec<&Path>>();
+
+    let result = pool.add_blockdevs(pool_uuid, &*pool_name, &blockdevs, tier);
+    let msg = match result.map(|bds| bds.changed()) {
+        Ok(Some(uuids)) => {
+            let return_value = uuids
+                .iter()
+                .map(|uuid| {
+                    // FIXME: To avoid this expect, modify add_blockdevs
+                    // so that it returns a mutable reference to each
+                    // blockdev created.
+                    create_dbus_blockdev(
+                        dbus_context,
+                        object_path.clone(),
+                        *uuid,
+                        pool.get_mut_blockdev(*uuid)
+                            .expect("just inserted by add_blockdevs")
+                            .1,
+                    )
+                })
+                .collect::<Vec<_>>();
+
+            return_message.append3((true, return_value), msg_code_ok(), msg_string_ok())
+        }
+        Ok(None) => return_message.append3(default_return, msg_code_ok(), msg_string_ok()),
+        Err(err) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&err);
+            return_message.append3(default_return, rc, rs)
+        }
+    };
+
+    Ok(vec![msg])
+}
+
+pub fn add_datadevs(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    add_blockdevs(m, BlockDevTier::Data)
+}
+
+pub fn add_cachedevs(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    add_blockdevs(m, BlockDevTier::Cache)
+}
+
+pub fn rename_pool(m: &MethodInfo<MTFn<TData>, TData>) -> MethodResult {
+    let message: &Message = m.msg;
+    let mut iter = message.iter_init();
+
+    let new_name: &str = get_next_arg(&mut iter, 0)?;
+
+    let dbus_context = m.tree.get_data();
+    let object_path = m.path.get_name();
+    let return_message = message.method_return();
+    let default_return = (false, uuid_to_string!(PoolUuid::nil()));
+
+    let pool_path = m
+        .tree
+        .get(object_path)
+        .expect("implicit argument must be in tree");
+    let pool_uuid = get_data!(pool_path; default_return; return_message).uuid;
+
+    let msg = match dbus_context
+        .engine
+        .borrow_mut()
+        .rename_pool(pool_uuid, new_name)
+    {
+        Ok(RenameAction::NoSource) => {
+            let error_message = format!("engine doesn't know about pool {}", pool_uuid);
+            let (rc, rs) = (DbusErrorEnum::INTERNAL_ERROR as u16, error_message);
+            return_message.append3(default_return, rc, rs)
+        }
+        Ok(RenameAction::Identity) => {
+            return_message.append3(default_return, msg_code_ok(), msg_string_ok())
+        }
+        Ok(RenameAction::Renamed(uuid)) => return_message.append3(
+            (true, uuid_to_string!(uuid)),
+            msg_code_ok(),
+            msg_string_ok(),
+        ),
+        Err(err) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&err);
+            return_message.append3(default_return, rc, rs)
+        }
+    };
+    Ok(vec![msg])
+}
