+++ src/dbus_api/pool/shared.rs
@@ -1,0 +1,35 @@
+use dbus::tree::{MTFn, Tree};
+
+use crate::{
+    dbus_api::types::TData,
+    engine::{Name, Pool, PoolUuid},
+};
+
+pub fn pool_operation<F, R>(
+    tree: &Tree<MTFn<TData>, TData>,
+    object_path: &dbus::Path<'static>,
+    closure: F,
+) -> Result<R, String>
+where
+    F: Fn((Name, PoolUuid, &dyn Pool)) -> Result<R, String>,
+    R: dbus::arg::Append,
+{
+    let dbus_context = tree.get_data();
+
+    let pool_path = tree
+        .get(object_path)
+        .expect("implicit argument must be in tree");
+
+    let pool_uuid = pool_path
+        .get_data()
+        .as_ref()
+        .ok_or_else(|| format!("no data for object path {}", object_path))?
+        .uuid;
+
+    let engine = dbus_context.engine.borrow();
+    let (pool_name, pool) = engine
+        .get_pool(pool_uuid)
+        .ok_or_else(|| format!("no pool corresponding to uuid {}", &pool_uuid))?;
+
+    closure((pool_name, pool_uuid, pool))
+}
